+++ src/venmo/shared/browser-detection.js
@@ -1,0 +1,17 @@
+'use strict';
+
+var isAndroid = require('@braintree/browser-detection/is-android');
+var isChrome = require('@braintree/browser-detection/is-chrome');
+var isIos = require('@braintree/browser-detection/is-ios');
+var isIosSafari = require('@braintree/browser-detection/is-ios-safari');
+var isSamsungBrowser = require('@braintree/browser-detection/is-samsung');
+var isMobileFirefox = require('@braintree/browser-detection/is-mobile-firefox');
+
+module.exports = {
+  isAndroid: isAndroid,
+  isChrome: isChrome,
+  isIos: isIos,
+  isIosSafari: isIosSafari,
+  isSamsungBrowser: isSamsungBrowser,
+  isMobileFirefox: isMobileFirefox
+};