+++ src/visa-checkout/visa-checkout.js
@@ -43,6 +43,7 @@ var cardTypeTransformMap = {
  * @property {string} nonce The payment method nonce.
  * @property {object} details Additional account details.
  * @property {string} details.cardType Type of card, ex: Visa, MasterCard.
+ * @property {string} details.lastFour Last four digits of card number.
  * @property {string} details.lastTwo Last two digits of card number.
  * @property {string} description A human-readable description.
  * @property {string} type The payment method type, always `VisaCheckoutCard`.